refactor: 토스트 관련 로직 훅으로 리팩토링
